Refactored: move data transformations to features
