+++ frontend-app/src/features/collection/types/index.ts
@@ -1,0 +1,5 @@
+export type CollectionDetail = {
+  bgImage: string;
+  parts: string;
+  name: string;
+};
