+++ frontend-app/src/features/crewDirectors/types/index.ts
@@ -1,0 +1,4 @@
+export type CrewDirectorDetails = {
+  name: string;
+  jobs: string;
+};
