+++ frontend-app/src/features/info/ui/Info/HomePageLink.tsx
