+++ frontend-app/src/features/info/ui/Info/ImdbLink.tsx
@@ -1,12 +1,12 @@
 import { Box, Text, Link } from "@chakra-ui/react"
 import { FaImdb } from "react-icons/fa";
 
-export function IMDBLink({ imdb_id }: { imdb_id?: string }) {
-    if (!imdb_id) {
+export function IMDBLink({ imdbLink }: { imdbLink?: string }) {
+    if (!imdbLink) {
         return <></>;
     }
     return <Box>
-        <Link href={`https://www.imdb.com/title/${imdb_id}`} target="_blank">
+        <Link href={`${imdbLink}`} target="_blank">
             <Text textStyle="3xl">
                 <FaImdb />
             </Text>
