+++ frontend-app/src/features/recommendation/types/index.ts
@@ -1,0 +1,9 @@
+export type RecommendationDetail = {
+  id: number;
+  title: string;
+  voteAverage: string;
+  backdrop: {
+    width: string;
+    imgSrc: string;
+  };
+};
