+++ frontend-app/src/features/social/types/index.ts
@@ -1,0 +1,5 @@
+export type ReviewDetails = {
+  id: string;
+  sliced_content: string;
+  fullContent: string;
+};
