+++ frontend-app/src/features/title/types/index.ts
@@ -1,0 +1,8 @@
+export type MovieTitle = {
+  duration: string;
+  countriesOrigin: string;
+  genres: string;
+  releaseDateLocale: string;
+  releaseDateYearLocale: string;
+  title: string;
+};
