+++ frontend-app/src/features/topBilledCast/types/index.ts
@@ -1,0 +1,7 @@
+export type CastImg = {
+  id: number;
+  width: string;
+  img: string;
+  name: string;
+  character: string;
+};
