+++ frontend-app/src/shared/ui/HorizontalItem.tsx
